basic temperature logging script
connects to wifi to set time from ntp server.
Wifi credentials are expected to be in a file wifi.cfg, having the
ssid on the first and the password on the second line. wifi.cfg is
ignored by git so that the password is not accidentally leaked.

the directory data is also ignored and should contain e.g. measurement results